# prov/shm: add shm signal to reduce lock contention